requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
google-cloud-documentai==2.25.0
google-cloud-storage==2.14.0
google-cloud-logging==3.8.0
google-cloud-secret-manager==2.16.4
google-auth>=2.27.0
google-api-core>=2.19.0
python-json-logger==2.0.7
python-magic==0.4.27
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
cachetools==5.3.2
python-Levenshtein==0.23.0
cryptography==41.0.7
oauthlib==3.2.2